HV:fix type related violations
1.Function return type inconsistent
2.cast on a constant value

V1->V2 add () to return type
V2->V3 keep the sbuf_get and sbuf_put return code

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/guest/instr_emul.c
@@ -2338,7 +2338,7 @@ int decode_instruction(struct vcpu *vcpu)
 		instr_check_gva(vcpu, emul_ctxt, cpu_mode);
 	}
 
-	return  emul_ctxt->vie.opsize;
+	return (int)(emul_ctxt->vie.opsize);
 }
 
 int emulate_instruction(struct vcpu *vcpu)
